+++ src/pages/content/project/Project.tsx
@@ -1,6 +1,6 @@
 import tw from 'tailwind-styled-components';
 import { useSelector } from 'react-redux';
-import ProjectCard from '../../../components/Cards/ProjectCard';
+import ProjectCard from '../../../components/cards/ProjectCard';
 import ProjectDetail from '../../detail/ProjectDetail';
 import { RootState } from '../../../store';
 import { TabsPropsT } from '../../../types/type';
